bonus pp equation fix?
unless I'm completely dumb, a "/" is a division...so 416.6667 / (1- 0.9994^N) doesn't seem to work as intended, with 25397 as N, it gives 416.6668(more than the max, which makes no sense), but with the other example that is supposed to give the half, 1168, it actually gives the double, 826.8571... . though, if you change the division for a multiplication, it works, it gives 209.965... , which is about the half. also, the have the closest to the half, it's 1155 and not 1168, apparently. I'm 99.9% sure of what I'm saying.